+++ src/component.cpp
@@ -27,9 +27,9 @@ Sorcery::Component::Component(): screen{""}, name{""}, x{0}, y{0}, w{0}, h{0}, s
 }
 
 Sorcery::Component::Component(std::string screen_, std::string name_, int x_, int y_, unsigned int w_, unsigned int h_,
-	float scale_, Enums::Internal::FontType font_, unsigned int size_, unsigned int colour_, std::string string_key_):
-	screen{screen_}, name{name_} , x{x_}, y{y_}, w{w_}, h{h_}, scale{scale_}, font{font_}, size{size_},
-	colour{colour_}, string_key{string_key_} {
+	float scale_, Enums::Internal::FontType font_, unsigned int size_, unsigned long long colour_,
+	std::string string_key_): screen{screen_}, name{name_} , x{x_}, y{y_}, w{w_}, h{h_}, scale{scale_}, font{font_},
+	size{size_}, colour{colour_}, string_key{string_key_} {
 }
 
 // Copy Constructor
